pyproject.toml
[tool.poetry]
name = "brocardapiproject"
version = "0.1.0"
description = ""
authors = ["flowerfrog <alexosenn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.3"
jsonschema = "^4.20.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
